+++ src/ui/RepositoryNavigator.h
@@ -11,11 +11,14 @@
 #include "git/Reference.h"
 #include "git/Repository.h"
 #include "host/GitHub.h"
+#include <QHash>
 #include <QPointer>
 #include <QWidget>
 #include <functional>
 
 class QComboBox;
+class QLabel;
+class QToolButton;
 class QTreeView;
 class RepositoryNavigatorModel;
 class RepoView;
@@ -27,14 +30,17 @@ public:
   using IssuesRequest = std::function<void(
       GitHub *, const QString &, const QString &,
       const GitHub::IssuesCallback &)>;
+  using Clock = std::function<qint64()>;
 
   explicit RepositoryNavigator(QWidget *parent = nullptr,
-                               const IssuesRequest &request = IssuesRequest());
+                               const IssuesRequest &request = IssuesRequest(),
+                               const Clock &clock = Clock());
 
   void setRepository(const git::Repository &repo);
   void setRepoView(RepoView *view);
   RepositoryNavigatorModel *model() const;
   QTreeView *view() const;
+  QTreeView *issuesView() const;
   QComboBox *issuesRemoteFilter() const;
   void setBodyFont(const QFont &font);
 
@@ -46,7 +52,10 @@ private:
   void showContextMenu(const QPoint &point);
   void discoverGitHubIssuesRepositories();
   void selectGitHubIssuesRepository(int index);
-  void requestGitHubIssues(bool refresh);
+  void requestGitHubIssues(bool force);
+  void applyGitHubIssuesCache(const QString &key);
+  void updateGitHubIssuesPanel();
+  QString currentGitHubIssuesKey() const;
   void openIssue(const QModelIndex &index);
   void showIssuesRepositoryMenu(const QModelIndex &index);
 
@@ -59,7 +68,22 @@ private:
     QPointer<GitHub> account;
   };
 
+  struct IssuesCacheEntry {
+    GitHub::Issues issues;
+    QString error;
+    qint64 lastSuccess = 0;
+    qint64 lastAttempt = 0;
+    qint64 retryAfter = 0;
+    int generation = 0;
+    bool hasValue = false;
+    bool inFlight = false;
+  };
+
   QTreeView *mView;
+  QWidget *mIssuesPanel;
+  QLabel *mIssuesTitle;
+  QToolButton *mIssuesRefresh;
+  QTreeView *mIssuesView;
   QComboBox *mIssuesRemoteFilter;
   RepositoryNavigatorModel *mModel;
   QPointer<RepoView> mRepoView;
@@ -70,8 +94,10 @@ private:
   git::Reference mReferenceBeforeReset;
   QList<QMetaObject::Connection> mRemoteConnections;
   QList<IssuesCandidate> mIssuesCandidates;
-  int mIssuesGeneration = 0;
+  QHash<QString, IssuesCacheEntry> mIssuesCache;
+  int mIssuesScrollBeforeReset = 0;
   IssuesRequest mIssuesRequest;
+  Clock mClock;
   QPointer<GitHub> mAnonymousGitHub;
 };